% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_blur.R
\name{gaussian_blur}
\alias{gaussian_blur}
\title{Gaussian blur}
\usage{
gaussian_blur(input_image_directory, output_image_directory, sigma)
}
\arguments{
\item{input_image_directory}{string, path to the input .png file}

\item{output_image_directory}{string, path to the output .png file}

\item{sigma}{double, the sd of the gaussian distribution}
}
\value{
png file at the output path
}
\description{
Gaussian blur
}
\examples{
\dontrun{
gaussian_blur("test.png", "test_output.png", sigma = 1)
}
}
