% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_filter.R
\name{laplacian_filter}
\alias{laplacian_filter}
\title{laplacian edge detecting}
\usage{
laplacian_filter(input_img, output_img)
}
\arguments{
\item{input_img}{string, path to the input .png file}

\item{output_img}{string, path to the output .png file}
}
\value{
png file at the output path
}
\description{
laplacian edge detecting
}
\examples{
laplacian_filter("input,png", "output.png")
}
