% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGB_manipulation.R
\name{RGB_manipulation}
\alias{RGB_manipulation}
\title{RGB manipulation}
\usage{
RGB_manipulation(input_path, output_path, R = 2L, G = 2L, B = 2L)
}
\arguments{
\item{input_path}{string, path to the input .png file}

\item{output_path}{string, path to the output .png file}

\item{R}{int, a weight value for red channel, equal to or greater than 0.}

\item{G}{int, a weight value for green channel, equal to or greater than 0.}

\item{B}{int, a weight value for blue channel, equal to or greater than 0.}
}
\value{
png file at the output path
}
\description{
RGB manipulation
}
\examples{
\dontrun{
RGB_manipulation("input.png", "output.png", R = 5L, G = 1L, B = 1L)
}
}
