% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_blur.R
\name{gaussian_blur}
\alias{gaussian_blur}
\title{Gaussian blur}
\usage{
gaussian_blur(input_img, output_img)
}
\arguments{
\item{input_img}{string, path to the input .png file}

\item{output_img}{string, path to the output .png file}
}
\value{
png file at the output path
}
\description{
Gaussian blur
}
\examples{
gaussian_blur("input,png", "output.png")
}
