% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGB_manipulation.R
\name{RGB_manipulation}
\alias{RGB_manipulation}
\title{RGB manipulation}
\usage{
RGB_manipulation(input_img, R = 1, G = 1, B = 1, output_img)
}
\arguments{
\item{input_img}{string, path to the input .png file}

\item{R}{int, a weight value, equal to or greater than 1, for red the filter}

\item{G}{int, a weight value, equal to or greater than 1, for green the filter}

\item{B}{int, a weight value, equal to or greater than 1, for blue the filter}

\item{output_img}{string, path to the output .png file}
}
\value{
png file at the output path
}
\description{
RGB manipulation
}
\examples{
RGB_manipulation("input.png", 1, 2, 3, "output.png")
}
